Support optional filter in yaml config
The documentation of the yaml config indicates that the filter list is optinal.
If the list was not provided, no tracks were returned as we didn't append the
track since there was no match. A default of name with ".*" is now set if not
other filter is present. The tests were updated to cover the new behaviour.
Additionally, the filter types were updated in the tests to match the supported
filters in provider.go.